First commit: create project versus and the first demo
Versus is a android game. It born as a my friend idea and I wanted to do this for him.
Therefore, I'm building that during Mobile Programming course, so maybe I will use that as exam project.

The main goal is simple: you have 20 secs to do the job. The job is to collocate a word in the right category.

Further features in the future.