scores: save and display scores
On the first login, a user is created on firestore. The key is his uid, and in the fields there is score_count (how many games he did).
After gameover, the new score will be saved on firestore. The key is uid_score-count (then update score_count).
In ScoresActivity the scores will be displayed, from higher.
However, this feature is on working